Switched to using the XML-RPC URL as a unique identifier for blogs in the datastore.
Users now cannot add multiple instances of the same blog. Previously users could add duplicates, which caused a number of issues. This is especially helpful if a user does "add all" for multisite/WP.com and then later tries it again with additional sites in the network.

Fixed #54. Deletion now works using the URL, so name conflicts are inconsequential.

Fixed #61. Secondary tiles now use the XML-RPC URL as the token identifier, so they can be added and deleted without issue.